Feat: 답글 작성자 확인 메서드를 Comment 내부로 이동

+++ src/main/java/threads/server/domain/comment/CommentService.java
@@ -5,16 +5,13 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import threads.server.application.exception.ForbiddenException;
 import threads.server.application.exception.NotFoundException;
-import threads.server.application.exception.UnauthorizedException;
 import threads.server.domain.comment.dto.*;
 import threads.server.domain.comment.repository.CommentRepository;
 import threads.server.domain.comment.repository.CommentRepositorySupport;
-import threads.server.domain.post.Post;
 import threads.server.domain.reply.dto.ReplyDto;
 import threads.server.domain.reply.repository.ReplyRepositorySupport;
-import threads.server.domain.user.User;
-import threads.server.domain.user.dto.UserDto;
 
 import java.util.List;
 
@@ -28,15 +25,15 @@ public class CommentService {
     private final ReplyRepositorySupport replyRepositorySupport;
 
     public CommentDto save(CreatingCommentDto commentDto) {
-        User user = new User(commentDto.getUserId());
-        Post post = new Post(commentDto.getPostId());
-        return toCommentDto(commentRepository.save(Comment.builder().user(user).post(post).content(commentDto.getContent()).build()));
+        return toCommentDto(commentRepository.save(Comment.builder().userId(commentDto.getUserId()).postId(commentDto.getPostId()).content(commentDto.getContent()).build()));
     }
 
     @Transactional
     public CommentDto update(UpdatingCommentDto commentDto) {
         Comment comment = commentRepository.findById(commentDto.getId()).orElseThrow(() -> new NotFoundException("댓글을 찾을 수 없습니다."));
-        authorizeUser(commentDto.getUserId(), comment.getUser().getId());
+        if(!comment.checkIfAuthor(commentDto.getUserId())) {
+            throw new ForbiddenException("권한이 없습니다.");
+        };
         comment.change(commentDto.getContent());
         commentRepository.save(comment);
         return toCommentDto(comment);
@@ -45,15 +42,10 @@ public class CommentService {
     @Transactional
     public void delete(Long commentId, Long userId) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new NotFoundException("댓글을 찾을 수 없습니다."));
-        authorizeUser(userId, comment.getUser().getId());
+        if(!comment.checkIfAuthor(userId)) {
+            throw new ForbiddenException("권한이 없습니다.");
+        };
         commentRepository.delete(comment);
-    }
-
-
-    private void authorizeUser(Long requestUserId, Long userIdFromComment) {
-        if (!requestUserId.equals(userIdFromComment)) {
-            throw new UnauthorizedException("권한이 없습니다.");
-        }
     }
 
     @Transactional
